Refactor google provider package name
