+++ src/main/java/gyro/provider/google/Copyable.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gyro.google;
+package gyro.provider.google;
 
 public interface Copyable<M> {
     void copyFrom(M model);
