+++ src/main/java/gyro/provider/google/GoogleResource.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gyro.google;
+package gyro.provider.google;
 
 import com.google.api.client.googleapis.services.json.AbstractGoogleJsonClient;
 import gyro.core.resource.Resource;
